+++ src/controller/RacingController.js
@@ -1,27 +1,23 @@
 const InputView = require('../view/InputView.js');
 const OutputView = require('../view/OutputView.js');
 const Validator = require('../utils/Validator.js');
-const RacingCar = require('../model/RacingCar.js');
-const Util = require('../utils/Util.js');
-const { UTIL_NUMBER } = require('../data/constants.js');
+const RacingGame = require('../model/RacingGame.js');
+
 const IO = require('../utils/IO.js');
 
 class RacingController {
-  #cars;
-  #tryCount;
+  #racingGame;
 
   constructor() {
-    this.#cars = [];
+    this.#racingGame = new RacingGame();
   }
 
   inputCarName() {
     InputView.readCarName((carNames) => {
       const carArr = carNames.split(',');
       Validator.validateCarNames(carArr);
+      this.#racingGame.setCarsName(carArr);
 
-      carArr.forEach((carName) => {
-        this.#cars.push(new RacingCar(carName));
-      });
       this.inputTryCount();
     });
   }
@@ -29,47 +25,17 @@ class RacingController {
   inputTryCount() {
     InputView.readTryCount((tryCount) => {
       Validator.validateTryCount(tryCount);
-      this.#tryCount = tryCount;
+      this.#racingGame.setTryCount(tryCount);
+
       OutputView.printWhiteSpace();
-      this.repeatProcess();
+
+      this.conductProcess();
     });
   }
 
-  assignRandom() {
-    this.#cars.forEach((car) => {
-      const randomValue = Util.randomValue();
-      this.judgeMove(randomValue, car);
-    });
-    OutputView.printMoveProcess(this.#cars);
-  }
-
-  judgeMove(randomValue, car) {
-    if (randomValue >= UTIL_NUMBER.CAR_MOVE_MINIMUM_NUMBER) {
-      car.move();
-    }
-  }
-
-  repeatProcess() {
-    OutputView.printMoveResult();
-    OutputView.printMoveProcess(this.#cars);
-    for (let i = 0; i < this.#tryCount; i++) {
-      this.assignRandom();
-    }
-    this.printWinner();
-  }
-
-  selectWinner() {
-    const carsMoveResults = new Map();
-    this.#cars.forEach((car) => {
-      carsMoveResults.set(car.getCarName(), car.getMoveCount());
-    });
-
-    const winners = Util.filterMaxObjects(carsMoveResults).map((obj) => obj[0]);
-    return winners;
-  }
-
-  printWinner() {
-    OutputView.printWinner(this.selectWinner());
+  conductProcess() {
+    this.#racingGame.repeatProcess();
+    this.#racingGame.printWinner();
     this.quitGame();
   }
